refactor(test): rename request.count.test.ts to request.test.ts
Per review feedback - count route tests belong in the main request test
file rather than a separate file per endpoint.
